ISSUE-22: Replace Koin with Hilt for dependency injection in FlixSample
