+++ Samples/Flix/build.gradle
@@ -19,6 +19,8 @@ plugins {
     id 'org.jetbrains.kotlin.android'
     id 'org.jetbrains.kotlin.plugin.serialization'
     id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin'
+    id 'kotlin-kapt'
+    id 'com.google.dagger.hilt.android'
 }
 
 apply plugin: 'dev.testify'
@@ -70,10 +72,16 @@ android {
     }
 }
 
+// Allow references to generated code
+kapt {
+    correctErrorTypes true
+}
+
 dependencies {
     // Core Android
     implementation 'androidx.core:core-ktx:1.10.1'
     implementation "androidx.lifecycle:lifecycle-runtime-compose:2.6.1"
+    implementation "androidx.appcompat:appcompat:1.3.1"
 
     // UI Layer
     def composeBom = platform('androidx.compose:compose-bom:2022.12.00')
@@ -87,8 +95,12 @@ dependencies {
     androidTestImplementation composeBom
 
     // Dependency injection
-    implementation 'io.insert-koin:koin-android:3.3.0'
-    implementation "io.insert-koin:koin-androidx-compose:3.3.0"
+    implementation "com.google.dagger:hilt-android:2.44"
+    kapt "com.google.dagger:hilt-compiler:2.44"
+
+    // Navigation
+    implementation "androidx.navigation:navigation-compose:2.6.0"
+    implementation "androidx.hilt:hilt-navigation-compose:1.0.0"
 
     implementation("org.jetbrains.kotlinx:kotlinx-datetime:0.4.0")
