+++ Samples/Flix/src/main/java/dev/testify/samples/flix/data/remote/tmdb/httpclient/TmdbHostnameProvider.kt
@@ -25,12 +25,8 @@
 
 package dev.testify.samples.flix.data.remote.tmdb.httpclient
 
-interface TmdbHostnameProvider {
-    fun provideHostname(): String
-}
+import javax.inject.Inject
 
-internal fun getTmdbHostnameProvider(): TmdbHostnameProvider {
-    return object : TmdbHostnameProvider {
-        override fun provideHostname() = "api.themoviedb.org"
-    }
+class TmdbHostnameProvider @Inject constructor() {
+    fun provideHostname() = "api.themoviedb.org"
 }
