+++ Samples/Flix/src/main/java/dev/testify/samples/flix/ui/moviedetails/MovieDetails.kt
@@ -32,9 +32,11 @@ import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import dev.testify.samples.flix.presentation.moviedetails.model.MovieDetailsPresentationModel
 import dev.testify.samples.flix.presentation.moviedetails.viewmodel.MovieDetailsViewModel
@@ -47,17 +49,18 @@ import dev.testify.samples.flix.ui.common.composeables.OverviewText
 import dev.testify.samples.flix.ui.common.composeables.PrimaryTitle
 import dev.testify.samples.flix.ui.common.composeables.SecondaryTitle
 import dev.testify.samples.flix.ui.common.renderer.ScreenState
-import org.koin.androidx.compose.koinViewModel
-import org.koin.core.parameter.parametersOf
 
 @Composable
 fun MovieDetailsScreen(
-    movieId: Int,
-    viewModel: MovieDetailsViewModel = koinViewModel { parametersOf(movieId) }
+    movieId: Int
 ) {
+    val viewModel = hiltViewModel<MovieDetailsViewModel>()
     val screenState by viewModel.screenState.collectAsStateWithLifecycle()
     ScreenState<MovieDetailsViewState>(screenState) { viewState ->
         MovieDetails(viewState)
+    }
+    LaunchedEffect(Unit) {
+        viewModel.initialize(movieId)
     }
 }
