+++ build.gradle
@@ -56,6 +56,7 @@ plugins {
     id 'org.jetbrains.dokka' version '1.8.10' apply false
     id 'org.jetbrains.kotlin.plugin.serialization' version '1.8.10'
     id 'com.google.android.libraries.mapsplatform.secrets-gradle-plugin' version '2.0.1' apply false
+    id 'com.google.dagger.hilt.android' version '2.44' apply false
 }
 
 task clean(type: Delete) {
